pyproject.toml
[tool.poetry]
name = "airflow-dags"
version = "0.1.1"
description = ""
authors = ["GersonRS"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
apache-airflow = {version = "^2.7.3", python = ">=3.10,<3.12"}
airflow-provider-mlflow = "^1.1.0"
airflow-provider-great-expectations = "^0.2.7"
apache-airflow-providers-apache-kafka = "^1.3.0"
apache-airflow-providers-apache-spark = "^4.5.0"
apache-airflow-providers-cncf-kubernetes = "^7.11.0"
apache-airflow-providers-papermill = "^3.6.0"
astro-sdk-python = {extras = ["postgres"], version = "^1.7.0"}
boto3 = "^1.34.2"
great-expectations = "^0.18.5"
h5py = "^3.10.0"
ipykernel = "^6.27.1"
matplotlib = "^3.8.2"
minio = "^7.2.0"
mlflow = "^2.9.2"
pandas = "^2.1.4"
pathlib = "^1.0.1"
scikit-image = "^0.22.0"
scikit-learn = "^1.3.2"
scipy = "^1.11.4"
seaborn = "^0.13.0"
statsmodels = "^0.14.1"
sympy = "^1.12"
hdbcli = "^2.19.20"
sqlalchemy-hana = "^1.1.1"
pyspark = "^3.5.0"
delta-spark = "^3.0.0"
setuptools = "^69.0.2"
wheel = "^0.42.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-docker-tools = "^3.1.3"
pytest-helpers-namespace = "^2021.12.29"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
